cleaning and arranging files
